feat: put day detection and content on homepage
[claude]
