+++ src/components/footer.tsx
@@ -7,7 +7,7 @@ export default function Footer() {
   return <footer className="sitefooter">
 
     <nav>
-      <Link href="/">Back to Calendar</Link>
+      <Link href="/">Calendar of all dates</Link>
     </nav>
 
     <p>Brought to you by <a href="http://kglw.net/?src=kglw.today&campaign=credits" target="_blank" rel="noreferrer">KGLW.net</a>!</p>
